Do not build on Windows in CI
Since valgrind is only available on POSIX-systems, but crucially not on
Windows as [documented here][docu]. Therefore it doesn't make much sense
to compile for Windows, when the underlying tool is not available. Not
supporting Windows might be beneficial for development, e.g. one can use
[`nix`] for POISX-functionality not available on Windows.
Note, that valgrind also support Android and similar, but this is not
relevant for this commit, as there is already no testing for those plat-
forms anyways.

[docu]: https://valgrind.org/info/platforms.html
[`nix`]: https://crates.io/crates/nix